Remove newrelic gem dependency
We haven't updated the gem for years and don't know whether it
still works with our current Ruby and Rails versions.

Besides, dependabot keeps opening pull requests to update. In theory we
could just ignore the dependabot pull requests for this dependency, but
unfortunately right now we can't add a dependabot config file because it
would open pull requests on forks as well.

Finally, there are other companies offering similar services for Rails
applications, and it's up to each CONSUL installation to decide which
one is better for them. We might add a self-hosted performance
monitoring tool in the future.

Since other CONSUL installations might be using Newrelic, and in general
we recommend adding an application monitoring tool, we're suggesting it
in the custom gemfile. In the name of neutrality, we're also adding
Sentry. We might add other services in the future.